Updated ants.py and bees.py with bug fixes
I fixed the bug when selecting New Ant/Bee in the menu and the bug that occasionally occurred when changing the rule string with time > 1000. I also made changes to bees.py, replacing the L's and R's with undirected arcs, and further highlighted the path that the bee is on in green.